+++ pycqed/instrument_drivers/physical_instruments/ZurichInstruments/wouter/UHFQuantumController.py
@@ -341,8 +341,9 @@ class UHFQC(Instrument):
         # The custom firmware will feed through the signals on Signal Input 1 to Signal Output 1 and Signal Input 2 to Signal Output 2
         # when the AWG is OFF. For most practical applications this is not really useful. We, therefore, disable the generation of
         # these signals on the output here.
-        self.sigouts_0_enables_3(0)
-        self.sigouts_1_enables_7(0)
+        if 0:   # FIXME: disabled because we get an error (20190702 WJV)
+            self.sigouts_0_enables_3(0)
+            self.sigouts_1_enables_7(0)
 
     def close(self) -> None:
         self._daq.disconnectDevice(self._device)
