+++ pycqed/tests/openql/test_clifford_rb_oql.py
@@ -48,3 +48,25 @@ class Test_char_rb_oql(unittest.TestCase):
             nr_cliffords=[2, 5, 11], nr_seeds=1,
             program_name='character_bench_int_CZ')
         self.assertEqual(p.name, 'character_bench_int_CZ')
+
+
+"""
+    Author:             Wouter Vlothuizen, QuTech
+    Purpose:            randomized benchmarking OpenQL tests for Qutech Central Controller
+    Notes:              requires OpenQL with CC backend support
+"""
+
+# NB: we just hijack the parent class to run the same tests
+class Test_cliff_rb_oql_CC(Test_cliff_rb_oql):
+    def setUp(self):
+        curdir = os.path.dirname(__file__)
+        self.config_fn = os.path.join(curdir, 'test_cfg_cc.json')
+        output_dir = os.path.join(curdir, 'test_output_cc')
+        ql.set_option('output_dir', output_dir)
+
+class Test_char_rb_oql_CC(Test_char_rb_oql):
+    def setUp(self):
+        curdir = os.path.dirname(__file__)
+        self.config_fn = os.path.join(curdir, 'test_cfg_cc.json')
+        output_dir = os.path.join(curdir, 'test_output_cc')
+        ql.set_option('output_dir', output_dir)
