+++ pycqed/tests/openql/test_single_qubit_oql.py
@@ -1,5 +1,6 @@
 import os
 import unittest
+import pytest
 import numpy as np
 
 try:
@@ -155,3 +156,24 @@ except ImportError as e:
         @unittest.skip('Missing dependency - ' + str(e))
         def test_fail(self):
             pass
+
+
+"""
+    Author:             Wouter Vlothuizen, QuTech
+    Purpose:            single qubit OpenQL tests for Qutech Central Controller
+    Notes:              requires OpenQL with CC backend support
+"""
+
+# NB: we just hijack the parent class to run the same tests
+class Test_single_qubit_seqs_CC(Test_single_qubit_seqs_CCL):
+    def setUp(self):
+        curdir = os.path.dirname(__file__)
+        self.config_fn = os.path.join(curdir, 'test_cfg_cc.json')
+        output_dir = os.path.join(curdir, 'test_output_cc')
+        ql.set_option('output_dir', output_dir)
+
+    def test_RTE(self):
+        pytest.skip("test_RTE() uses conditional gates, which are not implemented yet")
+
+    def test_fast_feedback_control(self):
+        pytest.skip("test_fast_feedback_control() uses conditional gates, which are not implemented yet")
